requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Bioinformatics
biopython>=1.81

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0

# Utils
tqdm>=4.65.0